+++ src/main/java/org/betonquest/betonquest/quest/condition/scoreboard/ScoreboardObjectiveConditionFactory.java
@@ -11,7 +11,7 @@ import org.betonquest.betonquest.quest.condition.PrimaryServerThreadPlayerCondit
 /**
  * Factory to create scoreboard conditions from {@link Instruction}s.
  */
-public class ScoreboardConditionFactory implements PlayerConditionFactory {
+public class ScoreboardObjectiveConditionFactory implements PlayerConditionFactory {
 
     /**
      * Data used for condition check on the primary server thread.
@@ -23,7 +23,7 @@ public class ScoreboardConditionFactory implements PlayerConditionFactory {
      *
      * @param data the data used for checking the condition on the main thread
      */
-    public ScoreboardConditionFactory(final PrimaryServerThreadData data) {
+    public ScoreboardObjectiveConditionFactory(final PrimaryServerThreadData data) {
         this.data = data;
     }
 
@@ -31,6 +31,6 @@ public class ScoreboardConditionFactory implements PlayerConditionFactory {
     public PlayerCondition parsePlayer(final Instruction instruction) throws InstructionParseException {
         final String objective = instruction.next();
         final VariableNumber count = instruction.getVarNum();
-        return new PrimaryServerThreadPlayerCondition(new ScoreboardCondition(objective, count), data);
+        return new PrimaryServerThreadPlayerCondition(new ScoreboardObjectiveCondition(objective, count), data);
     }
 }